feat: refactored hook - moved to Collection (#6483)
